backend/client_api/requirements.txt
alembic==1.5.8
apispec==4.4.0
attrs==20.3.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
Inject==4.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
mistune==0.8.4
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.11
webargs==8.0.0
Werkzeug==1.0.1
